backend/requirements.txt
# FastAPI
greenlet==2.0.2
anyio==3.6.2
click==8.1.3
fastapi==0.94.0
h11==0.14.0
idna==3.4
pydantic==1.10.6
sniffio==1.3.0
starlette==0.26.0.post1
uvicorn==0.21.0
psycopg2-binary

# pyrealsense2
pyrealsense2
numpy
opencv-python

# rasberry pi
RPi.GPIO
